Add a delay after receiving a character before sending
Receiving one character then sending can result in a lot of
unnecessary network traffic.  So add a small delay (two
character times) after receiving a character before sending
the data out on the wire.

This involved a lot of changes.  First we had to obtain the
bits per second and bits per character from the devio.  Then
we had to add a delay time instead of just doing an immediate
send of a character, unless the buffer was full or the delay
was meaningless.  Then we had to add calculation of the
proper delay.

Signed-off-by: Corey Minyard <cminyard@mvista.com>