Add support for triggering LEDs during serial traffic
This adds the possibility to flash an LED when serial traffic is
detected. The idea is to provide some kind of traffic indicator
like the flashing LEDs on e.g. network cards.

For this a tiny LED driver framework is added, with the first
driver 'sysfs'. This driver is funded on Linux's kernel LED
transient trigger.

It is possible to define different LEDs for both tx and rx directions,
and for one/multiple ports.

This changeset is targetting embedded devices with dedicated LEDs
running ser2net, however, you can also use the numlock/caps lock LEDs
of your PC keyboard to test.

Signed-off-by: Michael Heimpold <michael.heimpold@i2se.com>
Cc: Corey Minyard <minyard@acm.org>
Cc: Yegor Yefremov <yegorslists@googlemail.com>