Fix a bug that might cause IAC sends to be split
IACs are doubled in the main data, but we also send telnet commands
and it's in a separate buffer.  This could result in an IAC being
split and a telnet command being inserted in the middle of it.  So
rework the telnet command sending to wait until the current buffer
is sent before being sent.

Signed-off-by: Corey Minyard <cminyard@mvista.com>