Fix closeon and IAC doubling code
If you received more data from the network before you finished
sending it to the device, the closeon and IAC handling code
would process the wrong part of the buffer.  Change to process
from the last end to the current end of data.

Also, the copy of the data was copying one more byte than necessary,
and could theoretically result in an overrun.  Fix that, too.

Signed-off-by: Corey Minyard <cminyard@mvista.com>