netio: Fix a myriad of issues
Rework handling of shutdowns to not block, but instead call shutdown_done()
when finished.

Fix a bunch of issues with the UDP netio.  Lots of little locking issues,
don't create the sockets until the first startup call, remove the
caching of new connections that come in when disabled.

Lots of other little bug fixes.

Signed-off-by: Corey Minyard <cminyard@mvista.com>