Rework waiting generally
There were a few places that waiting for operations to complete
in certain cases where it made sense.  However, the waiting was
using pthread_cond(), and that doesn't integrate properly with
the selector code, as nothing will happen when pthread_cond()
is waiting.

Rework the wait code to run an event loop while waiting, so things
can run during the wait.

This means that all waiting must be done without any locks held.

To do this, the waiters needed to do a bunch of stuff, count how
many waits were necessary, and do a "bulk" wait for all of them
outside of any locks.  This require the wait code to take a count.

This fixes some broken/complicated locking in port reconfigure,
especially around when a reconfigure of a port is delayed when
the port is still in use in a reconfigure.

Signed-off-by: Corey Minyard <cminyard@mvista.com>