Fix telnet command issues
Fix a number of const correctness issues in the telnet and buffer
interface.

Also, the telnet cmd structure was being shared by all telnet connections,
which could result in a bunch of issues, since data is tracked there.
Make a copy of it in telnet_init().  This required returning an error
from telnet_init(), which required restructure elsewhere.

Signed-off-by: Corey Minyard <cminyard@mvista.com>