Clean up rfc2217 handling
There were a number of issues with rfc2217.

Fix the telnet handling of the will function to take will/do, and
modify the com port handling to properly return 0 for do commands.

The com port handling was enabled even if rfc2217 mode was disable,
disable it unless rfc2217 is enabled in the config and the other
side sent a will command for it.

The sergenio telnet handling wasn't sending a will for rfc2217.
Add that.

Delay the reporting of sergenio telnet open reporting until after
the response to the telnet com port "do" is received.

Signed-off-by: Corey Minyard <cminyard@mvista.com>