serialsim: Fix write draining issues
The handling of the transmission of data was done in the sender's
thread.  This means that when the sender shut down, it's thread
would stop running even if transmit data was pending to the receiver,
thus resulting in data not being sent when it should.

Move the transmission of data into the receiver's thread so that
even if the transmitter's thread stops, the receiver's thread will
still process the data.

Signed-off-by: Corey Minyard <cminyard@mvista.com>