genio: Start work to separate out the state machines from the handlers
There was a lot of duplicated code with state machine handling
of connections.  Create a generic state machine handler that
accepts a lower level connection and a filter handler that runs
the state machine.

Also add a file-descriptor based lower level connection.

Switch the TCP code over to use all of this.

Signed-off-by: Corey Minyard <cminyard@mvista.com>