genio: A few changes that are fairly intertwined
I did these together, and they were too hard to separate out.

Modify get_raddr to return an error and update addrlen.

Modify the way casting works for sergenio.  It now has pointers
to the parent object and child object instead of containment.
This is necessary for using the base genio for termios and telnet,
as containment doesn't work there.

Do a bunch of net->io renaming.

Signed-off-by: Corey Minyard <cminyard@mvista.com>