genio: Add a free refcount to avoid problems with multiple frees
There was a situation where a child was passed in to create an ll,
and there was a failure where there was a failure that caused the
ll to be freed, and thus the child to be freed.  Then the error
was returned and the caller freed the child.

Add a refcount to the genios to avoid this issue, so the free of
the child can be controlled.

Signed-off-by: Corey Minyard <cminyard@mvista.com>