genio: Add a generic genio acceptor over genio acceptor
This is generic code where you have an acceptor that sits on top of
another acceptor.  The user just provides a few simple functions
and the base code runs the state machine.

Convert genio_ssl over to use the generic code.

Remove the code that was set up to be converted to this in genio_base,
and put it in a different file.

Signed-off-by: Corey Minyard <cminyard@mvista.com>