genio: Fixes to get telnet server mode working
There were a number of issues with how telnet server mode
worked, the option numbers are different and there were a number
of issues setting up the proper environment for the telnet
object created in server mode.

Fixing the environment required adding a "finish" callback
to the genio over genio and a way to pass data from the
new child callback to the finish callback.  It's a little
more obfuscated than I would like, but it work.

In addition, add range checking to the telnet serial port
handling.

Signed-off-by: Corey Minyard <cminyard@mvista.com>