Remove the "max_read_size" and rework buffer handling
Modify all the genio layers to allow the read size to be specified
by an option.  Rework all the buffer sizes to be more sane for the
specific layers, UDP is 64K by default for the maximum packet size
so data doesn't get lost, and set SSL to it's maximum message block
so full packets will work properly.  Everything else is set to 1k.

Signed-off-by: Corey Minyard <cminyard@mvista.com>