gensio: Fix initial modemstate reporting
Recent changes removed the sergensio callbacks and thus the
call that set them.  Remove some leftover cruft from that,
and fix a missing modemstate report.

The code was using the callback set to know if the modemstate
could be reported.  Since it was removed, that no longer works.

So rework the gensio_base.c timer so it can be set even if the
gensio is not fully open, and set a timer if the modemstate
needs to be reported but cannot be reported yet because the
callbacks are not set.

Signed-off-by: Corey Minyard <cminyard@mvista.com>