Fix some bugs in kickolduser
Kickolduser had a couple of issues:  The wrong logic was used when
checking of something had slipped in before the user, and the new
port handling was done before the port was shut down, resulting
in a double-open of the port (and thus a failure).

For proper semantics, you really want the port to go through the
normal shutdown process so any close strings, open strings, etc
get sent.

So use the right logic, and delay the new port processing until
the port has been shut down.

Signed-off-by: Corey Minyard <cminyard@mvista.com>