Split out connect backs from remote addresses
Having connect backs in the remote address list was really a bad idea.
Create separate list and option for connect backs.

Note that the "!" will still work in the remaddr list and preserves the
semantics of it being both a remote address and connect back.  However,
it may go away in the future, it is no longer documented.

Signed-off-by: Corey Minyard <cminyard@mvista.com>