Fix issues with trace file handling
It was just broken for yaml.  Rework it to use trace-read/write/both
everywhere to make it easier to read.  Also add information about
special string handling to the man page, since it is used for trace
files.

Remove the rs485 info from the man page; gensio handles that now.
And move that handling to the deprecated section in rs485.c.

Signed-off-by: Corey Minyard <cminyard@mvista.com>