telnet: Fix telnet handling
After re-reading the telnet spec, it's about as clear as mud to me on
the WILL/WONT/DO/DONT handling.  The big thing is how to avoid an
infinite loop of these messages going back and forth.  The operative
part of RFC854 says:

      b. If a party receives what appears to be a request to enter some
      mode it is already in, the request should not be acknowledged.
      This non-response is essential to prevent endless loops in the
      negotiation.  It is required that a response be sent to requests
      for a change of mode -- even if the mode is not changed.

So you don't respond if the mode is not changed, but you must respond,
even if the mode is not changed.

Anyway, I have reworked all this handling to do the following:

* WILL and DO on commands with no handler will always result in a
  DONT or WONT.

* WONT and DONT on commands with no handler will be ignored.

* A WILL on an command that is set to not do will always result in a DONT.

* A WILL on a command that is set to do and has not had a DO sent
  already will result in a DO sent.

* A DO on an command that is set to not will will always result in a WONT.

* A DO on a command that is set to will and has not had a WILL sent
  already will result in a WILL sent.

So basically, commands with no handler will only respond DONT or WONT.
Commands with a handler will always send on DO/DONT/WILL/WONT on a
received message if nothing has been sent.  If something has already
been sent, only send something if the mode is changed.

Signed-off-by: Corey Minyard <cminyard@mvista.com>