Add yaml input and output to the controller interface
If the "yaml" command is executed, all output will be yaml and input is
not echoed.  This is yaml output mode.

If the controller sees "%YAML" in the input, it will switch to yaml
output mode and yaml input mode, where it expects commands to come in
the form:

  commmand:
    name: <cmd name>
    id: <id>
    parms: [ parm1 [, parm2 [...]]]

id will just be returned in the reponse if it is present.
parms are optional depending on the command.

Signed-off-by: Corey Minyard <cminyard@mvista.com>