Fix an issue with UDP port writes fail
If a UDP write fails, the read enable on the port wasn't being enabled.
The check to say that there was still data to write to the port was
returning that there was still data, but there obviously wasn't because
the UDP port failed.

The re-enable after the timeout also wasn't properly setting the state,
and there was a place that was clearing the write position when it
shouldn't have.

Also remove some debugging cruft that got left in earlier.

Signed-off-by: Corey Minyard <minyard@acm.org>