Fix an issue with dev to net lockup on a device closing
If a device exits while blocked on write output and there are mulitple
connections to the port, it is possible that if one connections closes
the system will stop sending data to the network.  This was due to the
write blocking not being handled on close.

Add a check on accepter connection close.

Signed-off-by: Corey Minyard <corey@minyard.net>